Fix arena "find opponent" crash on the bot-defender branch
The real bug behind "دکمه پیدا کردن حریف کار نمیکنه": when no real
player sits in the attacker's cup band, find_opponent() builds a bot
defender whose loot pool comes from random.randint(*arena_fake_loot_range()).
ARENA_FAKE_LOOT_PER_LEVEL is a float, so the range was (float, float) and
randint() raised TypeError. That's not a GameError, so arena_find_callback
didn't catch it — it had already answered the query (stopping the spinner)
and then died, so the button appeared to do nothing. It only reproduced
when matchmaking fell through to a bot, which is why it looked intermittent.

- int()-cast the bonus in arena_fake_loot_range so the range is integer.
- arena_find_callback now also catches unexpected exceptions, logs them,
  and shows an error alert instead of silently dying.

+++ game/constants.py
@@ -696,7 +696,10 @@ def arena_loot_cap(attacker_level: int) -> int:
 
 
 def arena_fake_loot_range(attacker_level: int) -> tuple[int, int]:
-    bonus = max(0, attacker_level) * ARENA_FAKE_LOOT_PER_LEVEL
+    # ARENA_FAKE_LOOT_PER_LEVEL is a float, so bonus is too — int() it, otherwise
+    # random.randint() (used on this range) raises TypeError on the fake-opponent
+    # branch and the whole "find opponent" flow silently dies.
+    bonus = int(max(0, attacker_level) * ARENA_FAKE_LOOT_PER_LEVEL)
     return ARENA_FAKE_LOOT_BASE[0] + bonus, ARENA_FAKE_LOOT_BASE[1] + bonus
